Abs to Relative path and Attribute Name Fix
