+++ pyupdate/main_cli.py
@@ -8,7 +8,7 @@ from pyupdate.utilities.build import BuildError
 def cli():
     """PyUpdate CLI"""
     parser = argparse.ArgumentParser(description='PyUpdate CLI')
-    parser.add_argument('-f', '--folder', help="Path to project folder", required=True)
+    parser.add_argument('-p', '--project', help="Path to project folder", required=True)
     parser.add_argument('-e', '--exclude', help="Exclude files and directories", nargs='+', default=[])
     args = parser.parse_args()
     
@@ -17,7 +17,7 @@ def cli():
         sys.exit(1)
     
     try:
-        builder = util.Builder(folder_path=args.folder, exclude_paths=args.exclude)
+        builder = util.Builder(project_path=args.project, exclude_paths=args.exclude)
         builder.build()
     except Exception as error:
         raise BuildError(error)
