+++ pyupdate/utilities/build.py
@@ -2,7 +2,7 @@
 
 import os
 import shutil
-from pyupdate.utilities import helper, hashing
+from pyupdate.utilities import helper, Hasher
 
 
 class BuildError(Exception):
@@ -35,7 +35,7 @@ class Builder:
     Builds a project into a pyupdate project
 
     Attributes:
-    folder_path: str
+    project_path: str
         Path to the project folder
     exclude_paths: list
         List of absolute paths to exclude from the hash database
@@ -44,8 +44,8 @@ class Builder:
     build() -> None
         Builds the project
     """
-    def __init__(self, folder_path: str, exclude_paths: list):
-        self.folder_path = folder_path
+    def __init__(self, project_path: str, exclude_paths: list):
+        self.project_path = project_path
         self.exclude_paths = exclude_paths
 
         self._pyudpdate_folder = None
@@ -78,28 +78,28 @@ class Builder:
     
     def _validate_paths(self):
         """Validates and set paths"""
-        if self.folder_path is None:
+        if self.project_path is None:
             raise BuildError('Folder path not set')
         if self.exclude_paths is None:
             raise BuildError('Exclude paths not set')
 
-        if not os.path.exists(self.folder_path):
-            raise FileNotFoundError(f'Folder "{self.folder_path}" does not exist')
-        if self.folder_path in self.exclude_paths:
+        if not os.path.exists(self.project_path):
+            raise FileNotFoundError(f'Folder "{self.project_path}" does not exist')
+        if self.project_path in self.exclude_paths:
             raise PathError(f'Folder path cannot be excluded')
         
         # Remove trailing slashes
-        if self.folder_path.endswith('/'):
-            self.folder_path = self.folder_path.rstrip('/')
-        if self.folder_path.endswith('\\'):
-            self.folder_path = self.folder_path.rstrip('\\')
+        if self.project_path.endswith('/'):
+            self.project_path = self.project_path.rstrip('/')
+        if self.project_path.endswith('\\'):
+            self.project_path = self.project_path.rstrip('\\')
         for i, path in enumerate(self.exclude_paths):
             if path.endswith('/'):
                 self.exclude_paths[i] = path.rstrip('/')
             if path.endswith('\\'):
                 self.exclude_paths[i] = path.rstrip('\\')
         
-        self._pyudpdate_folder = os.path.join(self.folder_path, '.pyupdate')
+        self._pyudpdate_folder = os.path.join(self.project_path, '.pyupdate')
         self._config_path = os.path.join(self._pyudpdate_folder, 'config.yaml')
         self._hash_db_path = os.path.join(self._pyudpdate_folder, 'hashes.db')
     
@@ -130,6 +130,8 @@ class Builder:
     def _create_hash_db(self):
         """Creates the hash database"""
         print(f'Creating hash database at "{self._hash_db_path}"')
+        hasher = Hasher(os.path.basename(self.project_path))
+
         # DEBUG Exclude paths need more testing
         excluded_paths = [self._pyudpdate_folder] + self.exclude_paths
-        hashing.create_hash_db(self.folder_path, self._hash_db_path, excluded_paths)
+        hasher.create_hash_db(self.project_path, self._hash_db_path, excluded_paths)
