--- pyupdate/utilities/hashing/__init__.py
@@ -1,1 +1,0 @@
-from .hasher import create_hash_db, compare_databases
