+++ setup.py
@@ -10,7 +10,7 @@ setup(
     author='Noah Blaszak',
     author_email='70231827+Trogiken@users.noreply.github.com',
     description='Python library that allows the version of a program to be updated.',
-    packages=['pyupdate', 'pyupdate.utilities', 'pyupdate.utilities.hashing'],
+    packages=['pyupdate', 'pyupdate.utilities'],
     package_data={'pyupdate.utilities': ['default.yml', 'comments.yml']},
     install_requires=requirements,
     entry_points={
